# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyPrent Accounting",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assets": {
      "get": {
        "tags": [
          "Asset Registry"
        ],
        "summary": "List Assets",
        "operationId": "asset_registry_list_assets",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/gnucash/accounts": {
      "get": {
        "tags": [
          "Settings and Backup"
        ],
        "summary": "Read Gnucash Accounts",
        "operationId": "settings_and_backup_read_gnucash_accounts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GnuCashAccountResponse"
                  },
                  "type": "array",
                  "title": "Response Settings And Backup Read Gnucash Accounts"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetListResponse": {
        "properties": {
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrganizationResponse"
            },
            "type": "array",
            "title": "Organizations"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/PropertyResponse"
            },
            "type": "array",
            "title": "Properties"
          },
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/BuildingResponse"
            },
            "type": "array",
            "title": "Buildings"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitResponse"
            },
            "type": "array",
            "title": "Units"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/RoomResponse"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "organizations",
          "properties",
          "buildings",
          "units",
          "rooms"
        ],
        "title": "AssetListResponse"
      },
...
      "BuildingResponse": {
        "properties": {
          "property_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "year_built": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Built"
          },
          "street": {
            "type": "string",
            "title": "Street"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name"
          },
          "unit_count": {
            "type": "integer",
            "title": "Unit Count",
            "default": 0
          },
          "room_count": {
            "type": "integer",
            "title": "Room Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "street",
          "city",
          "postal_code",
          "id"
        ],
        "title": "BuildingResponse"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "reason"
        ],
        "title": "ErrorDetail",
        "description": "A stable machine code and a readable explanation."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The common HTTP error envelope."
      },
...
      "GnuCashAccountResponse": {
        "properties": {
          "guid": {
            "type": "string",
            "title": "Guid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "parent_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Guid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "guid",
          "name",
          "full_name",
          "parent_guid"
        ],
        "title": "GnuCashAccountResponse"
      },
...
      "OrganizationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "OrganizationResponse"
      },
...
      "PropertyResponse": {
        "properties": {
          "organization_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "street": {
            "type": "string",
            "title": "Street"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "building_count": {
            "type": "integer",
            "title": "Building Count",
            "default": 0
          },
          "unit_count": {
            "type": "integer",
            "title": "Unit Count",
            "default": 0
          },
          "room_count": {
            "type": "integer",
            "title": "Room Count",
            "default": 0
          },
          "expense_count": {
            "type": "integer",
            "title": "Expense Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "organization_id",
          "name",
          "street",
          "city",
          "postal_code",
          "id"
        ],
        "title": "PropertyResponse"
      },
...
      "RoomResponse": {
        "properties": {
          "unit_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Unit Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "area_sqm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Sqm"
          },
          "area_share_percent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Share Percent"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "unit_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Label"
          },
          "building_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Id"
          },
          "property_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Id"
          },
          "building_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Name"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit_id",
          "label",
          "id"
        ],
        "title": "RoomResponse"
      },
...
      "UnitResponse": {
        "properties": {
          "building_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "area_sqm": {
            "type": "string",
            "title": "Area Sqm"
          },
          "mea_percent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mea Percent"
          },
          "room_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Room Count"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "property_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Id"
          },
          "building_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Name"
          },
          "property_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Name"
          },
          "actual_room_count": {
            "type": "integer",
            "title": "Actual Room Count",
            "default": 0
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "area_sqm",
          "room_count",
          "id"
        ],
        "title": "UnitResponse"
      },
...
    }
  }
}